+++ app/src/main/java/com/puc/pi3_es_2024_t24/manager/ManagerActivity.kt
@@ -1,9 +1,12 @@
 package com.puc.pi3_es_2024_t24.manager
 
+import android.content.Intent
+import android.nfc.NfcAdapter
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
+import androidx.navigation.fragment.NavHostFragment
 import com.puc.pi3_es_2024_t24.R
 
 class ManagerActivity : AppCompatActivity() {
@@ -20,4 +23,19 @@ class ManagerActivity : AppCompatActivity() {
         // retorna a pilha anterior
         return navController.navigateUp() || super.onSupportNavigateUp()
     }
+
+    override fun onNewIntent(intent: Intent?) {
+        super.onNewIntent(intent)
+        intent?.let {
+            if (NfcAdapter.ACTION_NDEF_DISCOVERED == it.action ||
+                NfcAdapter.ACTION_TECH_DISCOVERED == it.action ||
+                NfcAdapter.ACTION_TAG_DISCOVERED == it.action) {
+                val navHostFragment =
+                    supportFragmentManager.findFragmentById(R.id.navHostFragmentContainerView) as NavHostFragment
+                val confirmLockerFragment = navHostFragment.childFragmentManager.fragments[0] as ConfirmLockerFragment
+                confirmLockerFragment.newIntent(it)
+            }
+        }
+    }
+
 }
