+++ app/src/main/java/com/puc/pi3_es_2024_t24/models/NfcTag.kt
@@ -1,6 +1,5 @@
 package com.puc.pi3_es_2024_t24.models
 
 data class NfcTag(
-    var method: String,
-    val messages: String
+    var method: String
 )
